Fix/ modifypage 체중, 키, birthdate 추가 완료 체중, 키는 유효성검사 및 기존데이터 입력완료
