+++ healthmanagement_back/src/main/java/com/portfolio/healthmanagement/dto/account/ModifyUserInfoReqDto.java
@@ -1,5 +1,7 @@
 package com.portfolio.healthmanagement.dto.account;
 
+import java.sql.Date;
+
 import lombok.Data;
 
 @Data
@@ -7,6 +9,9 @@ public class ModifyUserInfoReqDto {
 	private int userId;
 	private String phone;
 	private String email;
-	private String address;
+	private int weight;
+	private int height;
+	private Date birthdate;
+	
 
 }
